python/pyproject.toml
# Licensed to the Apache Software Foundation (ASF) under one or more
# contributor license agreements.  See the NOTICE file distributed with
# this work for additional information regarding copyright ownership.
# The ASF licenses this file to You under the Apache License, Version 2.0
# (the "License"); you may not use this file except in compliance with
# the License.  You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[tool.poetry]
name = "rocketmq"
version = "0.1.0rc1"
description = "RocketMQ Python Client"
authors = ["RocketMQ Authors"]
readme = "README.md"
packages = [{ include = "rocketmq" }]

[tool.poetry.dependencies]
python = "^3.7"
grpcio = "^1.50.0"
grpcio-tools = "^1.50.0"
protobuf = "^4.21.1"
certifi = "^2023.5.7"
apscheduler = "^3.10.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.07"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.*  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''
